pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "devon-traffic"
version = "0.1.0"
description = "Synthetic GPS/OD traffic dashboard for Devon, Torbay and Plymouth (FastHTML + Plotly)."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "plai" }]
dependencies = [
    "python-fasthtml>=0.12",
    "plotly>=5.20",
    "pandas>=2.2",
    "numpy>=1.26",
    "click>=8.1",
    "uvicorn>=0.30",
    "pyarrow>=15",
]

[project.optional-dependencies]
test = ["playwright>=1.45", "pillow>=10", "pytest>=8"]
docs = ["pillow>=10", "reportlab>=4.0"]

[project.scripts]
devon-traffic = "devon_traffic.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"
